requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0 